+++ WebContent/WEB-INF/dispatcher-servlet.xml
@@ -10,7 +10,7 @@
                            http://www.springframework.org/schema/mvc
                            http://www.springframework.org/schema/mvc/spring-mvc.xsd ">
 
-   <context:component-scan base-package="services" />
+   <context:component-scan base-package="rms.services" />
 
    <bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
       <property name="prefix" value="/WEB-INF/views/" />
