+++ WebContent/WEB-INF/views/dashboard.jsp
@@ -17,8 +17,9 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
 
   
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.0/umd/popper.min.js"></script>
+
+ 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.0/umd/popper.min.js"></script>
     	
 	<spring:url value="/resources/css/bootstrap.min.css" var="bootMin"/>
 	<spring:url value="/resources/css/dashboard.css" var="DashboardCSS"/>
@@ -27,100 +28,13 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
 	<link rel="stylesheet" href="${DashboardCSS}"/>
 	
 	
-	<spring:url value="/Js" var="JS"/>
+  	<spring:url value="/resources/js" var="JS"/>
 	
 	<script src="${JS}/jquery.js"></script>
 	<script src="${JS}/bootstrap.js"></script>
 	
 
-  <style>
-    /* Remove the navbar's default margin-bottom and rounded borders */ 
-   
-      
-      
-      
-      .container {
-   padding:0px;
-    
-       
-}
-      
- 
-      .container, #footer{
-          background-color:white;
-     
-        
-      }
-      .navbar {
-      margin-bottom: 0;
-      border-radius: 0;
-     background-color:white;
-       
-    }
-    
-      
-  
-      
-    /* Set height of the grid so .sidenav can be 100% (adjust as needed) */
-    .row.content {
-        height: 550px;
-        
-      
-      }
-    
-    /* Set gray background color and 100% height */
-    .sidenav {
-      padding-top: 20px;
-      background-color:rgb(0,51,0);
-      height: 550px;
-    }
-      
-      
-      div.row.content{
-          background-image: url("resources/images/green.jpg");
-          
-      }
-    
-    /* Set black background color, white text and some padding */
-    footer {
-     color:rgb(0,51,0);
-       background-color: white;
-      height:60px;
-    
-      padding: 0px;
-        position: absolute;
-        width: inherit;
-    }
-    
-    /* On small screens, set height to 'auto' for sidenav and grid */
-    @media screen and (max-width: 767px) {
-      .sidenav {
-        height: auto;
-        padding: 15px;
-      }
-      .row.content {
-          height:auto;
-        } 
-    }
-      
-      
-          @media screen and (max-width: 767px) {
-      .card {
-        max-height: 100px;
-        max-width:100px;
-        padding: 15px;
-      }
-      }
-      
-      
-      #bod{
-          
-          background-color:#eee9e9	;
-        /*background-image: url("green.jpg"); /* fallback for old browsers */
 
-      }
-      
-  </style>
 </head>
 <body id="bod">
     <div class="container">
@@ -167,18 +81,18 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
     
   
 <div class="container text-center" >    
-  <div class="row content">
+  <div class="row Content" style="background-image: url(/Phoenix_Resource_Management_System/resources/images/green.jpg);">
     <div class="col-sm-1 sidenav">
         <span id="linkside">
        
         
-            <div class="panel-group">
+     <div class="panel-group">
 
        
         <li class="dropdown">
           <a href="#" class="dropdown-toggle" id="burger" style="font-color:white;" data-toggle="dropdown"><span class="glyphicon glyphicon-menu-hamburger" style="font-size:50px; text-align:center; color:white;"></span></a>
           <ul class="dropdown-menu">
-              <br>
+            
             <li><a href="#">Book <span class="glyphicon glyphicon-book pull-right"></span></a></li>
             <li class="divider"></li>
              
@@ -197,7 +111,7 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
         </li>
         
    
-            </div>
+      </div>
             
             
             
@@ -213,7 +127,7 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
  <div class="well well-lg" style="background-color:rgb(0,50,0); color:white;"><h1>   Resource Booker </h1> </div> -->
         <div class="help" style="background-color:rgb(0,50,0); border-radius:25px;">
           <center>  
-            <img src="resources/images/rms.png" alt="rmsLogo" style=width:300px;">
+            <img src="resources/images/rms.png" alt="rmsLogo" style="width:300px;">
             
             </center>
  </div>
@@ -223,7 +137,7 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
             
        
             
-   <div class="row">
+   <div class="row" >
        
   <div class="column">
       
@@ -234,7 +148,7 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
       <p>  <span class="glyphicon glyphicon-book" style="font-size:150px;"></span></p>
     </div>
   </div>
-<center>
+
   <div class="column">
 <br>
   </div> 
@@ -264,16 +178,17 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
            
      
   
-      </center> 
+     
    
            
 </div>
-
-
 <footer class="footer" id="footer"">
-  <p>© 2018 Syntel, Inc </p>
+ <center> <p>© 2018 Syntel, Inc </p></center>
                                           
 </footer>   
+
+</div>
+
 
 </body>
 </html>
