+++ WebContent/WEB-INF/views/loginfailed.jsp
@@ -47,7 +47,8 @@ org.springframework.web.context.support.WebApplicationContextUtils"%>
 	      <p class="message">Already registered? <a href="#">Sign In</a></p>
 	    </form>
 	      
-	 <form action="loginOnUserName" method="post">
+	 <form action="dashboard" method="post">
+	 <span>Please enter a valid username/password !</span>
 	  <div class="input-group">
 	    <span class="input-group-addon"><i class="glyphicon glyphicon-user"></i></span>
 	    <input id="email" type="text" class="form-control" name="userName" placeholder="userName" value='${user.getUserName()}'>
