+++ WebContent/resources/css/dashboard.css
@@ -39,7 +39,7 @@
       
       
       div.row.content{
-          background-image: url("resources/images/green.jpg");
+          background: url("/Phoenix_Resource_Management_System/resources/images/green.jpg");
           
       }
     
@@ -48,7 +48,7 @@
      color:rgb(0,51,0);
        background-color: white;
       height:60px;
-    
+      width: 100%;
       padding: 0px;
         position: absolute;
         width: inherit;
