--- src/services/Services.java
@@ -1,19 +1,0 @@
-package services;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-
-@Controller
-public class Services {
-	
-	@RequestMapping(value="/")
-	public String homeScreen() {
-		return "login";
-	}
-	
-	@RequestMapping(value="/dashboard")
-	public String dashBoard() {
-		return "dashboard";
-	}
-	
-}
